log-analysis-project/requirements.txt
python-evtx>=0.7; python_version >= "3.6"
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
plotly>=5.0
dash>=2.0
reportlab>=3.6
matplotlib>=3.3
python-dateutil>=2.8
pytest>=6.0
requests>=2.0